* Create the Spectroscopy base class for spectroscopy experiments.
